sync groups with what ubiquity installs would ACTUALLY produce
this is actually somewhat fishy as I think part of the groups with
ubiquity come from adduser defaults rather than being hardcoded anywhere
in the installer in general.
probably the best approximation we can do for now in calamares though.

the incorrect lp group specifically is causing questionable read access to
resource which ordinarily aren't readable and screws up snap confinement
as it allows access() on resources to succeed but later get denied by
apparmor when tried to actual read it (which then screws up libcups/qt)