add hacks to hopefully fix GPT partioning
cala's autopartition created a new Partition instance with the initial
flag being Esp already. As a result kpmcore's setpartflagsjob didn't
attempt to set Esp anymore as it thought it is set already

the autopartition also set tried to set the bootflag, which made no sense
to sfdisk as boot means nothing on gpt. esp is equal to boot.